Tegra: Fix the delay loop used during SC7 exit
This patch fixes the delay loop used to wake up the BPMP during SC7 exit.
The earlier loop would fail just when the timer was about to wrap-around
(e.g. when TEGRA_TMRUS_BASE is 0xfffffffe, the target value becomes 0,
which would cause the loop to exit before it's expiry).

Change-Id: I7dd66d475fde97a18ca00dfc96ce610294263151
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Signed-off-by: Furquan Shaikh <furquan@google.com>