Tegra: implement per-SoC validate_power_state() handler
The validate_power_state() handler checks the power_state for a valid afflvl
and state id. Although the afflvl check is common, the state ids are implementation
defined.

This patch moves the handler to the tegra/soc folder to allow each SoC to validate
the power_state for supported parameters.

Change-Id: Ia037052a01c432d86e5b4b2095ba2966e2e1c9c2
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>