rockchip: rk3399: Don't wait for vblank in M0 for ddrfreq
This removes waiting for vblank on the M0 during ddrfreq transitions.
That will now be done in the kernel to allow scheduling to be done on
the CPU core that changes the ddr frequency. Waiting for vblank in
the M0 would have the CPU core that waits on the M0 spin looping for
up to 16ms (1 frame for the display).

Change-Id: I11bbaa44eddc98c22a4b513f5c4ce0888bcfdc65
Signed-off-by: Derek Basehore <dbasehore@chromium.org>