rockchip: rk3399: Move DQS drive strength setting to M0
This moves the setting of the DQS drive strength to the M0 to minimize
the impact on DDR transactions. We need to have the DQS drive strength
changed for data training, which is triggered by the M0, but it also
needs to be changed back when data training is finished.

BUG=chrome-os-partner:62513
TEST=boot kevin
stressapptest -M 1500 -s 100000
check for no errors

Change-Id: I4e61201a87ea5d55fb605aa20b2510e82575a38f
Signed-off-by: Derek Basehore <dbasehore@chromium.org>